pyproject.toml
[project]
name = "openclaw-py"
version = "0.1.0"
description = "Minimal Python port of OpenClaw Agent Harness — all intelligence, no channel bloat"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "tiktoken>=0.7.0",
    "numpy>=1.26.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
openclaw-py = "openclaw.repl:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
